refactor(admin): split content detail shell
Why: keep the dynamic route focused on param handling while moving D1 loading and UI sections into named modules that are easier to review and reuse.

Verify: pnpm exec prettier --check 'apps/admin/src/app/content/[id]/page.tsx' 'apps/admin/src/app/content/[id]/content-detail-data.ts' 'apps/admin/src/app/content/[id]/content-detail-shell.tsx'; git diff --check -- 'apps/admin/src/app/content/[id]/page.tsx' 'apps/admin/src/app/content/[id]/content-detail-data.ts' 'apps/admin/src/app/content/[id]/content-detail-shell.tsx'; pnpm --filter @anipotts/admin typecheck; pnpm --filter @anipotts/admin lint; pnpm --filter @anipotts/admin build; pnpm validate
