+++ apps/admin/src/app/content/[id]/content-detail-shell.tsx
@@ -1,0 +1,190 @@
+import Link from "next/link";
+import StatusActions from "./status-actions";
+import EditableContent from "./editable-content";
+import AtomManager from "./atom-manager";
+import ButtondownCard from "./buttondown-card";
+import type { ContentDetailData } from "./content-detail-data";
+import { SERIES_COLORS, STATUS_COLORS, PLATFORM_ABBREV } from "@/lib/constants";
+
+function ContentTopBar({ thought }: Pick<ContentDetailData, "thought">) {
+  const status = thought.status || "draft";
+
+  return (
+    <div className="shrink-0 border-b border-zinc-800 px-6 py-3 flex items-center justify-between bg-zinc-900/50">
+      <div className="flex items-center gap-3 min-w-0">
+        <Link
+          href="/content"
+          className="text-zinc-500 hover:text-zinc-300 transition-colors shrink-0"
+        >
+          <svg
+            className="w-4 h-4"
+            fill="none"
+            viewBox="0 0 24 24"
+            stroke="currentColor"
+            strokeWidth={2}
+          >
+            <path
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              d="M15.75 19.5L8.25 12l7.5-7.5"
+            />
+          </svg>
+        </Link>
+        <h2 className="text-sm font-medium truncate">{thought.title}</h2>
+        <span
+          className={`text-[11px] px-2 py-0.5 rounded-full shrink-0 ${STATUS_COLORS[status] || STATUS_COLORS.idea}`}
+        >
+          {status}
+        </span>
+        {thought.series_type && (
+          <span
+            className={`text-[11px] px-2 py-0.5 rounded-full shrink-0 ${SERIES_COLORS[thought.series_type] || ""}`}
+          >
+            {thought.series_type}
+          </span>
+        )}
+      </div>
+      <a
+        href={`https://anipotts.com/writing/${thought.slug}`}
+        target="_blank"
+        rel="noopener noreferrer"
+        className="text-xs text-zinc-500 hover:text-zinc-300 transition-colors flex items-center gap-1 shrink-0"
+      >
+        View on site
+        <svg
+          className="w-3 h-3"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke="currentColor"
+          strokeWidth={2}
+        >
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            d="M13.5 6H5.25A2.25 2.25 0 003 8.25v10.5A2.25 2.25 0 005.25 21h10.5A2.25 2.25 0 0018 18.75V10.5m-10.5 6L21 3m0 0h-5.25M21 3v5.25"
+          />
+        </svg>
+      </a>
+    </div>
+  );
+}
+
+function ContentDetails({ thought }: Pick<ContentDetailData, "thought">) {
+  return (
+    <section>
+      <h4 className="admin-label">Details</h4>
+      <div className="text-xs text-zinc-400 space-y-1.5">
+        {thought.content_type && (
+          <div>
+            Type: <span className="text-zinc-300">{thought.content_type}</span>
+          </div>
+        )}
+        <div>
+          Created:{" "}
+          <span className="text-zinc-300">
+            {new Date(thought.created_at).toLocaleDateString()}
+          </span>
+        </div>
+        {thought.published_at && (
+          <div>
+            Published:{" "}
+            <span className="text-zinc-300">
+              {new Date(thought.published_at).toLocaleDateString()}
+            </span>
+          </div>
+        )}
+        <div>
+          Views: <span className="text-zinc-300">{thought.views}</span>
+        </div>
+        {thought.artifact_url && (
+          <div>
+            Artifact:{" "}
+            <a
+              href={thought.artifact_url}
+              target="_blank"
+              rel="noopener noreferrer"
+              className="text-indigo-400 hover:text-indigo-300 break-all"
+            >
+              {thought.artifact_type || "link"}
+            </a>
+          </div>
+        )}
+      </div>
+
+      {thought.platforms_posted && thought.platforms_posted.length > 0 && (
+        <div className="flex items-center gap-1 mt-2">
+          <span className="text-[11px] text-zinc-500">Posted:</span>
+          {thought.platforms_posted.map((p) => (
+            <span
+              key={p}
+              className="text-[10px] px-1.5 py-0.5 bg-green-500/10 text-green-400 rounded"
+            >
+              {PLATFORM_ABBREV[p] || p}
+            </span>
+          ))}
+        </div>
+      )}
+    </section>
+  );
+}
+
+function ContentSidebar({ thought, atoms }: ContentDetailData) {
+  const status = thought.status || "draft";
+
+  return (
+    <div className="w-72 shrink-0 border-l border-zinc-800/60 overflow-y-auto admin-scroll">
+      <div className="p-3 space-y-4">
+        <section>
+          <h4 className="admin-label">Status</h4>
+          <StatusActions id={thought.id} currentStatus={status} />
+        </section>
+
+        <ContentDetails thought={thought} />
+
+        <section>
+          <h4 className="admin-label">Email</h4>
+          <ButtondownCard
+            contentId={thought.id}
+            initialEmailId={thought.buttondown_email_id ?? undefined}
+          />
+        </section>
+
+        <section>
+          <h4 className="admin-label">Distribution</h4>
+          <AtomManager contentId={thought.id} atoms={atoms} />
+        </section>
+
+        <section className="pb-4">
+          <Link
+            href={`/record/${thought.id}`}
+            className="text-xs px-3 py-2 bg-zinc-800 hover:bg-zinc-700 rounded-md text-zinc-400 transition-colors block text-center"
+          >
+            Teleprompter
+          </Link>
+        </section>
+      </div>
+    </div>
+  );
+}
+
+export function ContentDetailShell({ thought, atoms }: ContentDetailData) {
+  return (
+    <div className="h-screen flex flex-col">
+      <ContentTopBar thought={thought} />
+
+      <div className="flex-1 flex overflow-hidden">
+        <div className="flex-1 overflow-y-auto admin-scroll p-6">
+          <EditableContent
+            id={thought.id}
+            title={thought.title}
+            summary={thought.summary || ""}
+            content={thought.content || ""}
+            tags={Array.isArray(thought.tags) ? thought.tags : []}
+          />
+        </div>
+
+        <ContentSidebar thought={thought} atoms={atoms} />
+      </div>
+    </div>
+  );
+}
